Refactor webpack config into common module
