+++ webpack-example.config.js
@@ -1,27 +1,11 @@
 const webpack = require('webpack');
 const path = require('path');
-const conf = require('./tasks/config');
+const common  = require('./webpack-common');
 
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const UglifyJSPlugin = require('uglifyjs-webpack-plugin');
 
-const { lstatSync, readdirSync } = require('fs')
-const { join } = require('path')
-
-const isDirectory = source => lstatSync(source).isDirectory()
-const getDirectories = source =>
-  readdirSync(source).map(name => join(source, name)).filter(isDirectory)
-
-const subDirs = getDirectories('./examples');
-
-const entry = {};
-for (let i = 0, ii = subDirs.length; i < ii; ++i) {
-  const name = subDirs[i].split(path.sep).pop();
-  if (conf.skip.indexOf(name) === -1) {
-    entry[name] = [];
-    entry[name].push(`.${path.sep}${subDirs[i]}${path.sep}app.js`);
-  }
-}
+const entry = common.getEntries(false);
 
 const config = {
   resolve: {
